backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium
requests
python-dotenv
ddddocr
